voice-tutor/requirements.txt
# Azure AI Voice Live API
azure-ai-voicelive[aiohttp]>=1.0.0

# Audio processing
pyaudio>=0.2.13

# Environment variables
python-dotenv>=1.0.0

# Azure authentication
azure-identity>=1.15.0

# Additional dependencies for audio handling
numpy>=1.24.0
